refactor the product model (separate the products from it's variants) for simplicity
