+++ src/types/product.types.ts
@@ -1,20 +1,10 @@
 import { Document, Model, Types } from 'mongoose';
 import { z } from 'zod';
+
 import {
   createProductSchema,
   updateProductSchema,
-} from '../validation/products.validation';
-
-export type VariantDocument = {
-  price: number;
-  oldPrice?: number;
-  stock: number;
-  sku?: string;
-  //   images: [string];
-  status: string;
-  size: string;
-  color: string;
-};
+} from '../validation/product.validation';
 
 export type ProductDocument = Document & {
   _id: Types.ObjectId;
@@ -23,7 +13,6 @@ export type ProductDocument = Document & {
   brand?: string;
   categories: [Types.ObjectId];
   description: string;
-  variants: [VariantDocument];
   //   averageRating: number;
   //   reviewsNumber: number;
 };
